Prueba_API/requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
openai
scikit-learn
numpy